+++ package.json
@@ -26,6 +26,7 @@
     "react-dom": "^17.0.2",
     "react-icons": "^3.0.0",
     "react-redux": "^8.0.2",
+    "react-switch": "^7.0.0",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "reselect": "^4.1.6",
